requirements.txt
flask==3.0.2
flask-cors==4.0.0
flask-limiter==3.5.0
requests==2.31.0
python-dotenv==1.0.1
# Optional: Uncomment if you want to use Gemini API
# google-generativeai>=0.3.0 